+++ dtool_lookup_client_dependency_graph_plugin/__init__.py
@@ -1,4 +1,4 @@
-"""dtool_lookup_client package."""
+"""dtool_lookup_client_dependency_graph_plugin package."""
 
 from datetime import date, datetime
 
@@ -15,6 +15,7 @@ import dtoolcore.utils
 import dtool_config.cli
 import dtool_lookup_api
 
+from .utils import graph_to_dot, build_graph_from_dataset_list
 
 logger = logging.getLogger(__name__)
 
@@ -50,14 +51,20 @@ DTOOL_LOOKUP_SERVER_TOKEN_KEY = "DTOOL_LOOKUP_SERVER_TOKEN"
 @click.option("--dependency-keys", 'dependency_keys', default=None)
 @click.option("-p", "--page-number", 'page_number', default=1, type=int)
 @click.option("-s", "--page-size", 'page_size', default=10, type=int)
-def graph(uuid, dependency_keys, page_number, page_size):
+@click.option("--dot",  is_flag=True, show_default=True, default=False, help="Print graph in dot format.")
+def graph(uuid, dependency_keys, page_number, page_size, dot):
     """Print the URIs associated with a UUID in the lookup server."""
     pagination = {}
     r = dtool_lookup_api.graph(uuid, dependency_keys=dependency_keys, page_number=page_number, page_size=page_size, pagination=pagination)
     logger.info(f"Pagination information: {pagination}")
-    formatted_json = json.dumps(r, indent=2)
-    colorful_json = pygments.highlight(
-        formatted_json,
-        pygments.lexers.JsonLexer(),
-        pygments.formatters.TerminalFormatter())
-    click.secho(colorful_json, nl=False)
+    if dot:
+        graph = build_graph_from_dataset_list(r)
+        txt = graph_to_dot(graph)
+        click.secho(txt, nl=False)
+    else:
+        formatted_json = json.dumps(r, indent=2)
+        colorful_json = pygments.highlight(
+            formatted_json,
+            pygments.lexers.JsonLexer(),
+            pygments.formatters.TerminalFormatter())
+        click.secho(colorful_json, nl=False)